add MapViewConnected component exports
